+++ src/App/components/Panel/index.test.js
@@ -115,13 +115,13 @@ describe("Component: Panel - ", () => {
         expect(wrapper.html()).toContain("panel-half-padding");
     });
 
-    it("renders with class 'panel-dark-mode' when prop darkMode is true", () => {
+    it("renders with class 'panel-dark' when prop darkMode is true", () => {
         const wrapper = shallow(
             <Panel darkMode={true} />
         );
 
         expect(wrapper).toMatchSnapshot();
-        expect(wrapper.html()).toContain("panel-dark-mode");
+        expect(wrapper.html()).toContain("panel-dark");
     });
 
     it("prop type is an enum with set values", () => {